t/xml/texts-1.dat

#data escaped
\uD800\uD801\uDFF0\uDFFF
#errors
1;1;text outside of root element;;;m
1;4;no root element;;;m
#document escaped
| "\uD800\uD801\uDFF0\uDFFF"

#data
]]>
#errors
1;1;unmatched mse;;;m
1;1;text outside of root element;;;m
1;3;no root element;;;m
#document
| "]]>"

#data
]]]>
#errors
1;2;unmatched mse;;;m
1;1;text outside of root element;;;m
1;4;no root element;;;m
#document
| "]]]>"

#data
]]]]>
#errors
1;3;unmatched mse;;;m
1;1;text outside of root element;;;m
1;5;no root element;;;m
#document
| "]]]]>"

#data
]]]]]>
#errors
1;4;unmatched mse;;;m
1;1;text outside of root element;;;m
1;6;no root element;;;m
#document
| "]]]]]>"

#data
aa]]>bb
#errors
1;3;unmatched mse;;;m
1;1;text outside of root element;;;m
1;7;no root element;;;m
#document
| "aa]]>bb"

#data
<![CDATA[]]>]]>
#errors
1;13;unmatched mse;;;m
1;1;cdata outside of root element;;;m
1;15;no root element;;;m
#document
| "]]>"

#data
<a/>]]>
#errors
1;5;unmatched mse;;;m
1;5;text outside of root element;;;m
#document
| <{} a>
| "]]>"

#data
<!---->]]>
#errors
1;8;unmatched mse;;;m
1;8;text outside of root element;;;m
1;10;no root element;;;m
#document
| <!--  -->
| "]]>"

#data
<>
#errors
1;1;bare stago;;;m
1;1;text outside of root element;;;m
1;2;no root element;;;m
#document
| "<>"

#data
<p><></p>
#errors
1;4;bare stago;;;m
#document
| <{} p>
|   "<>"

#data escaped
\u0000<p></p>
#errors
1;1;text outside of root element;;;m
1;1;NULL;;;m
1;1;NULL;;;m
#document escaped
| "\uFFFD"
| <{} p>

#data escaped
a\u0000b<p></p>
#errors
1;1;text outside of root element;;;m
1;2;NULL;;;m
1;2;NULL;;;m
#document escaped
| "a\uFFFDb"
| <{} p>

#data escaped
<!DOCTYPE p>\u0000<p></p>
#errors
1;13;text outside of root element;;;m
1;13;NULL;;;m
1;13;NULL;;;m
#document escaped
| <!DOCTYPE p>
| "\uFFFD"
| <{} p>

#data escaped
<!DOCTYPE p>a\u0000b<p></p>
#errors
1;13;text outside of root element;;;m
1;14;NULL;;;m
1;14;NULL;;;m
#document escaped
| <!DOCTYPE p>
| "a\uFFFDb"
| <{} p>

#data escaped
<p>\u0000</p>
#errors
1;4;NULL;;;m
1;4;NULL;;;m
#document escaped
| <{} p>
|   "\uFFFD"

#data escaped
<p>a\u0000b</p>
#errors
1;5;NULL;;;m
1;5;NULL;;;m
#document escaped
| <{} p>
|   "a\uFFFDb"

#data escaped
<p></p>\u0000
#errors
1;8;NULL;;;m
1;8;NULL;;;m
1;8;text outside of root element;;;m
#document escaped
| <{} p>
| "\uFFFD"

#data escaped
<p></p>a\u0000b
#errors
1;8;text outside of root element;;;m
1;9;NULL;;;m
1;9;NULL;;;m
#document escaped
| <{} p>
| "a\uFFFDb"
